refactor: move mocks to it's own directory
